First GUI implementing
added a login/create user functionality with the use for txt file. I know we may use a DB for this later...Also got a barebones main interactive screen. I envision this screen to be the place where the user can navigate to the different parts of our program (features of our program).