Fixes error and login message bug
This commit changed the way that the GUI class handles the login and database connection error. Previously errors connecting to the database was only logged in the console because only a specific exception was caught. The new changes make the connection an error message and a faild login a regular error.