Update to height and weight button, user info menu, user class.
When editing user information the after user inputs height and weight
information the system gives feedback in the form of a pop up.
The user's sex is now stored as a single character string instead
of a char... this change was necessary for being able to get data 
from the database. Because of the order everything is loaded I cannot
figure out how to get the buttons to display the appropriate values
when the user goes to alter their data...