GUI screens are made only after user is logged in. Buttons in alter user settings menu show current user data.
The GUI was making all of the screens before the program could actually
receive user information. To fix this issue I called the functions to 
create each menu after the session could be made. I then made it so 
that the alter user info function gives the buttons the user's current 
information.