Back end for set nutrient goals now works
Connected the front end to the back end for nutrient goals. To accomplish 
this I had to edit the name of the buttons. As it is now the buttons 
were created all at once and had an event listener added to them while 
they were being made. The event listener has an anonymous function called. 
The anonymous function calls show input dialog, takes in one argument 
the button's name. I used that to update the database. If the user does 
not put anything the database does nothing. If the user does input a value 
the database is updated using the button's name and the value the user 
passes in.